Fix heap overflows, XMSS key reuse, LMS verification, and KEM oracles
Addresses an external security audit covering XMSS, LMS, SNOVA, CROSS
and three KEMs, plus three further defects found while verifying it.

Critical:

- SNOVA and CROSS declared a max_signature_size smaller than what their
  serialisers emit, so sign() wrote past a correctly sized caller buffer
  on every call.  SNOVA sizes now derive from (n, l); CROSS sizes are
  worst-case bounds for the encoding it actually produces.  Both packers
  refuse rather than exceed the buffer.

- NTRU and NTRU Prime took their public sizes from the respective
  specifications while the code uses a larger encoding, overflowing pk,
  sk and ct in keygen and encaps.  Sizes now derive from the real layout
  in one place.  This also fixes decapsulation deriving the wrong shared
  secret, which followed from the truncated key material.

- ntru_poly_inv_mod_q sized abandoned scratch arrays NTRU_MAX_N but
  indexed them at [n], a stack overrun on NTRU-HPS-4096-821.  The dead
  arrays are removed.

- XMSS derived an identical WOTS+ one-time key for every leaf: the WOTS+
  routines re-set the address type, which zeroes the type-specific
  fields and wiped the leaf index the caller had just set.  Signatures
  still round-tripped, so round-trip tests could not see it.

High:

- LMS signatures never verified.  The authentication path was derived
  from the secret seed instead of the Merkle tree, and the verifier's
  left/right child order was inverted.  The path now comes from the real
  tree, shared with root computation.

- The XMSS WOTS+ checksum dropped its most significant nibble, leaving
  the final chain constant at zero.

- HQC, BIKE and FrodoKEM returned the decapsulation validity bit,
  exposing a plaintext-checking oracle through the public API.  They now
  always return PQC_OK with branchless ciphertext comparison; the same
  branch is removed from NTRU Prime.

Medium:

- Six verify() implementations ignored signature_len and read out of
  bounds on truncated input.  pqc_sig_verify() now rejects over-long
  lengths before dispatch, and each implementation validates its own.

- XMSS per-node hash keys and bitmasks collapsed to one pair per tree
  level because the key/bitmask selector aliased the tree index.  It now
  uses the RFC 8391 keyAndMask field at offset 28.

- pqc_randombytes() failures were ignored at 14 sites, deriving keys and
  nonces from uninitialised memory.  All sites propagate the failure.

- Signed integer overflow in FN-DSA ntru_solve.c now uses unsigned
  accumulation.

LMS-SHA256-H20 and H25 keygen now fail cleanly instead of substituting a
hash of the secret seed for the Merkle root.  CROSS, NTRU and NTRU Prime
key and signature sizes changed; keys from earlier builds are not
compatible.

Adds tests/unit/test_security_audit.c, which asserts the violated
invariant for each defect rather than round-trip success.  52 checks
pass; the full suite is green.

Reported-by: Conner (Spartan8806) <https://github.com/spartan8806>

+++ src/core/sig/fndsa/ntru_solve.c
@@ -255,7 +255,10 @@ fndsa_solve_ntru(unsigned logn,
     for (lv = logn; lv > 0; lv--) {
         size_t cn = (size_t)1 << lv;
         size_t chn = cn >> 1;
-        int64_t *tmp64;
+        /* Accumulate in unsigned: the field-norm products routinely exceed
+         * int64_t range, and signed overflow is undefined behaviour even
+         * where the generated code happens to wrap as intended. */
+        uint64_t *tmp64;
 
         fi32[lv - 1] = (int32_t *)malloc(chn * sizeof(int32_t));
         gi32[lv - 1] = (int32_t *)malloc(chn * sizeof(int32_t));
@@ -265,13 +268,13 @@ fndsa_solve_ntru(unsigned logn,
             { rc = -1; goto done; }
 
         /* int32_t field norms (truncated -- Babai tolerates this). */
-        tmp64 = (int64_t *)calloc(chn, sizeof(int64_t));
+        tmp64 = (uint64_t *)calloc(chn, sizeof(uint64_t));
         if (!tmp64) { rc = -1; goto done; }
         for (i = 0; i < chn; i++) {
             size_t j;
             for (j = 0; j < chn; j++) {
                 size_t idx = i + j;
-                int64_t p = (int64_t)fi32[lv][2 * i] * (int64_t)fi32[lv][2 * j];
+                uint64_t p = (uint64_t)(int64_t)fi32[lv][2 * i] * (uint64_t)(int64_t)fi32[lv][2 * j];
                 if (idx >= chn) { idx -= chn; tmp64[idx] -= p; }
                 else { tmp64[idx] += p; }
             }
@@ -280,21 +283,21 @@ fndsa_solve_ntru(unsigned logn,
             size_t j;
             for (j = 0; j < chn; j++) {
                 size_t idx = i + j + 1;
-                int64_t p = (int64_t)fi32[lv][2 * i + 1] * (int64_t)fi32[lv][2 * j + 1];
+                uint64_t p = (uint64_t)(int64_t)fi32[lv][2 * i + 1] * (uint64_t)(int64_t)fi32[lv][2 * j + 1];
                 if (idx >= chn) { idx -= chn; tmp64[idx] += p; }
                 else { tmp64[idx] -= p; }
             }
         }
-        for (i = 0; i < chn; i++) fi32[lv - 1][i] = (int32_t)tmp64[i];
+        for (i = 0; i < chn; i++) fi32[lv - 1][i] = (int32_t)(uint32_t)tmp64[i];
         free(tmp64);
 
-        tmp64 = (int64_t *)calloc(chn, sizeof(int64_t));
+        tmp64 = (uint64_t *)calloc(chn, sizeof(uint64_t));
         if (!tmp64) { rc = -1; goto done; }
         for (i = 0; i < chn; i++) {
             size_t j;
             for (j = 0; j < chn; j++) {
                 size_t idx = i + j;
-                int64_t p = (int64_t)gi32[lv][2 * i] * (int64_t)gi32[lv][2 * j];
+                uint64_t p = (uint64_t)(int64_t)gi32[lv][2 * i] * (uint64_t)(int64_t)gi32[lv][2 * j];
                 if (idx >= chn) { idx -= chn; tmp64[idx] -= p; }
                 else { tmp64[idx] += p; }
             }
@@ -303,12 +306,12 @@ fndsa_solve_ntru(unsigned logn,
             size_t j;
             for (j = 0; j < chn; j++) {
                 size_t idx = i + j + 1;
-                int64_t p = (int64_t)gi32[lv][2 * i + 1] * (int64_t)gi32[lv][2 * j + 1];
+                uint64_t p = (uint64_t)(int64_t)gi32[lv][2 * i + 1] * (uint64_t)(int64_t)gi32[lv][2 * j + 1];
                 if (idx >= chn) { idx -= chn; tmp64[idx] += p; }
                 else { tmp64[idx] -= p; }
             }
         }
-        for (i = 0; i < chn; i++) gi32[lv - 1][i] = (int32_t)tmp64[i];
+        for (i = 0; i < chn; i++) gi32[lv - 1][i] = (int32_t)(uint32_t)tmp64[i];
         free(tmp64);
 
         /* Modular field norms (exact mod P). */
